Server/backend/accounts/requirements.txt
django
PyMysql
wheel
mysqlclient
sorl-thumbnail
channels 
channels-redis
daphne
python-decouple
Pillow
oauth2-provider
django-oauth-toolkit 
djangorestframework
hypercorn
django-cors-headers
uvicorn
drf-yasg
drf-yasg[validation]
django 
djangorestframework 
djangorestframework-simplejwt 
celery 
redis 
mysqlclient 
scipy

